requirements.txt
requests
beautifulsoup4
lxml
faiss-cpu
numpy
sentence-transformers
groq
python-dotenv